Testing has distinctions within software/hardware, refactored to make it make sense
