+++ src/webots_pkg/test/hardware/test_pre_launch_cf_lighthouse_configuration.py
@@ -33,66 +33,7 @@ from cflib.crazyflie.syncCrazyflie import SyncCrazyflie
 from cflib.utils import uri_helper
 
 
-def test_ReadLHMem(radio_uri = "radio://0/80/2M/E7E7E7E7E7"):
-    uri = uri_helper.uri_from_env(radio_uri)
-    cflib.crtp.init_drivers()
 
-    print()
-    print('------------------')
-    print("INITIAL MEMORY")
-    print()
-    temp_LHMem = ReadLHMem(radio_uri)
-    assert temp_LHMem.geo_data[0] is not None
-    assert temp_LHMem.calib_data[0] is not None
+# TODO: Add tests to confirm results from cf that current lh configuration is correct. 
 
 
-def test_WriteLHMem(radio_uri = "radio://0/80/2M/E7E7E7E7E7"):    # 
-    uri = uri_helper.uri_from_env(radio_uri)
-    cflib.crtp.init_drivers()
-
-    # Create a LighthouseBsGeometry object
-    bs1geo = LighthouseBsGeometry()
-    bs1geo.origin = [0.0, 0.0, 1.0] # Best guess on wheere the base station is on tb
-    bs1geo.rotation_matrix = [
-        [0.0, -1.0, 0.0],
-        [1.0, 0.0, 0.0],
-        [0.0, 0.0, 7.0],
-    ]
-    bs1geo.valid = True
-
-    # Read the current values
-    print()
-    print('------------------')
-    print("INITIAL MEMORY FOR WRITING LH CONFIG")
-    temp_cf_lh_memory = ReadLHMem(radio_uri)
-    print()
-    calib_data = temp_cf_lh_memory.calib_data[0]
-
-    # Write geo memory to crazyflie
-    geo_dict = {0: bs1geo}
-    calib_dict = {0: calib_data}
-    WriteLHGeoMem(radio_uri, geo_dict)
-
-    # Read new values
-    print()
-    print('------------------')
-    print("NEW MEMORY AFTER WRITING LH CONFIG")
-    print()
-    new_memory = ReadLHMem(radio_uri)
-    assert new_memory.geo_data[0].origin == bs1geo.origin
-
-
-def test_get_cf_pose():
-    get_cf_pose()
-    
-
-def test_get_cf_pose_according_to_lh():
-    get_cf_pose_according_to_lh()
-
-
-
-if __name__ == "__main__":
-    # test_ReadLHMem()
-    # test_WriteLHMem()
-    # test_get_cf_pose()
-    test_get_cf_pose_according_to_lh()
