+++ src/webots_pkg/test/software/test_lighthouse_node.py
@@ -1,0 +1,10 @@
+import os
+import sys
+import pytest
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+from webots_pkg.lighthouse_node import LighthouseNode
+
+
+
+#TODO_tests: Write testing for this file
